Move all "Panels" to the `default.html` Layout, for now.
This commit prepares the site to continue to function as a multi-page
site instead of a low-fidelity version of a Single Page Application
(SPA) by moving the SPA-like behaviors into the layout portion of the
template instead of the single `index.html` file.

It should not actually cause a change in behavior or functionality. This
is just refactor intended as preparation for a larger change, later.